Move cell setup code to an extension of a StyledTableViewCell
